+++ dev/user.clj
@@ -4,8 +4,8 @@
 
 (css/set-css!
  "https://unpkg.com/computer-modern@0.1.2/cmu-serif.css"
- "https://unpkg.com/mafs@0.15.2/core.css"
- "https://unpkg.com/mafs@0.15.2/font.css")
+ "https://unpkg.com/mafs@0.17.0/core.css"
+ "https://unpkg.com/mafs@0.17.0/font.css")
 
 (def index
   "dev/mafs/notebook.clj")